bug windows

+++ Generator/include/CGAL/random_convex_hull_in_disc_2.h
@@ -196,7 +196,7 @@ void random_convex_hull_in_disc_2(std::size_t n, double radius, std::list<typena
            (simulated_points < n));  // initialization such that 0 in P_n
 
   std::size_t T = n;
-  if (!fast) T = static_cast<std::size_t>(std::floor(n / std::pow(std::log(n), 2)));
+  if (!fast) T = static_cast<std::size_t>(std::floor(n / std::pow(std::log(static_cast<double>(n)), 2)));
   
   while (simulated_points < n) {
     // l is a list coming from a convex hull operation. we are moving the
